feat: add admin and user roles for fine grained method level authorization (#4)

+++ src/main/java/com/example/ordermanagementsystem/services/UserServiceImpl.java
@@ -6,10 +6,17 @@ import com.example.ordermanagementsystem.config.EntityMapper;
 import com.example.ordermanagementsystem.entity.User;
 import com.example.ordermanagementsystem.exception.CustomGraphQLException;
 import com.example.ordermanagementsystem.input.CreateUserInput;
+import com.example.ordermanagementsystem.input.LoginUserInput;
 import com.example.ordermanagementsystem.input.UpdateUserInput;
 import com.example.ordermanagementsystem.repository.UserRepository;
+import com.example.ordermanagementsystem.security.TokenGenerator;
+import graphql.GraphQLException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -23,6 +30,8 @@ import java.util.stream.Collectors;
 public class UserServiceImpl implements UserService{
     private final UserRepository userRepository;
     private final EntityMapper entityMapper;
+    private final TokenGenerator tokenGenerator;
+    private final PasswordEncoder passwordEncoder;
     @Override
     public UserPayload createUser(CreateUserInput input) {
         Optional<User> existingUser = userRepository.findUserByEmail(input.getEmail());
@@ -31,21 +40,25 @@ public class UserServiceImpl implements UserService{
         }
         User userToBeSaved = User.builder().email(input.getEmail())
                 .name(input.getName())
-                .password(input.getPassword())
+                .password(passwordEncoder.encode(input.getPassword()))
+                .role(input.getRole())
                 .build();
         User savedUser = userRepository.save(userToBeSaved);
         return entityMapper.userToUserPayload(savedUser);
     }
 
+    @Secured("IS_AUTHENTICATED_FULLY")
     @Override
     public UserPayload user(Long id) {
-        Optional<User> existingUser = userRepository.findUserById(id);
-        if (existingUser.isEmpty()){
-            throw new CustomGraphQLException(String.format("User with id %s does not exist", id), 404);
+        String emailFromToken = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Optional<User> existingUser = userRepository.findUserByEmail(emailFromToken);
+        if (!existingUser.get().getId().equals(id)){
+            throw new CustomGraphQLException("Not authorized to view this user's details", 401);
         }
         return entityMapper.userToUserPayload(existingUser.get());
     }
 
+    @Secured("ROLE_ADMIN")
     @Override
     public List<UserPayload> users() {
         return userRepository.findAll().stream().map(
@@ -53,28 +66,51 @@ public class UserServiceImpl implements UserService{
                         id(user.getId())
                         .email(user.getEmail())
                         .name(user.getName())
+                        .role(user.getRole())
                         .build()
         ).collect(Collectors.toList());
     }
 
+    @Secured("ROLE_NORMAL")
     @Override
     public UserPayload updateUser(UpdateUserInput payload){
-        Optional<User> existingUser = userRepository.findUserById(payload.getId());
+        String emailFromToken = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Optional<User> existingUser = userRepository.findUserByEmail(emailFromToken);
         if (existingUser.isEmpty()){
             throw new CustomGraphQLException(String.format("User with id %s does not exist", payload.getId()), 404);
         }
-        entityMapper.updateFields(existingUser.get(), payload);
-        var updatedUser = userRepository.save(existingUser.get());
+        if (!existingUser.get().getId().equals(payload.getId())){
+            throw new CustomGraphQLException("Not authorized to view this user's details", 401);
+        }
+        var updatedUser = userRepository.save(entityMapper.updateUserInputToUser(existingUser.get(), payload));
         return entityMapper.userToUserPayload(updatedUser);
         }
 
+    @Secured("ROLE_NORMAL")
     @Override
     public GenericMessage deleteUser(Long id) {
-        Optional<User> existingUser = userRepository.findUserById(id);
+        String emailFromToken = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Optional<User> existingUser = userRepository.findUserByEmail(emailFromToken);
         if (existingUser.isEmpty()){
             throw new CustomGraphQLException(String.format("User with id %s does not exist", id), 404);
         }
+        if (!existingUser.get().getId().equals(id)){
+            throw new CustomGraphQLException("Not authorized to view this user's details", 401);
+        }
         userRepository.deleteById(id);
         return new GenericMessage(String.format("User with id %s has successfully been deleted", id));
+    }
+
+    @Override
+    public String loginUser(LoginUserInput input) {
+        var existingUser = userRepository.findUserByEmail(input.getEmail());
+        if (existingUser.isEmpty()){
+            throw new CustomGraphQLException(String.format("User with email %s does not exist", input.getEmail()), 404);
+        }
+
+        if (!passwordEncoder.matches(input.getPassword(), existingUser.get().getPassword())){
+            throw new CustomGraphQLException("Invalid credentials", 401);
+        }
+        return tokenGenerator.build(existingUser.get().getEmail(), existingUser.get().getRole());
     }
 }
